Make asClassDeclaration fluent; align DeepPathUtilGenerator
Change RGeneratedJavaClass#asClassDeclaration to return a dual TargetLanguage
representation that renders via the fluent CodeWriter API and, through a reverse
bridge, still appends to the legacy Xtend template machinery. FunctionGenerator
now renders the default class via asClassDeclaration() again (so defaultClass is
an RGeneratedJavaClass<?>), and the not-yet-migrated generators that embed it in
Xtend templates keep working unchanged.

Extract the reverse bridge TargetStringConcatenationCodeWriter from JavaExpression
into a reusable public class (the counterpart of CodeWriterTargetStringConcatenation).

For consistency with FunctionGenerator#classBody, DeepPathUtilGenerator#renderClass
now returns a CodeRenderer instead of being a void method taking a CodeWriter.
